Separate out the hw context specific code to a new file
